[librvth] RefFile::size(): BLKGETSIZE64 requires an int64_t for the return value.
Otherwise, we end up smashing the stack.

This fixes a regression from commit 34171a2231f8b07011246329481b5b76df914ca7.
([librvth] Rewrote RefFile as a C++ class.)

+++ src/librvth/RefFile.cpp
@@ -294,7 +294,7 @@ int64_t RefFile::size(void)
 #elif defined(__linux__)
 		// Linux version.
 		// Reference: http://www.microhowto.info/howto/get_the_size_of_a_linux_block_special_device_in_c.html
-		int ret = -1;
+		int64_t ret = -1;
 		if (ioctl(fileno(m_file), BLKGETSIZE64, &ret) == 0) {
 			// Size obtained successfully.
 			return ret;
